code and word firepad working
